requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
Pillow==10.4.0
requests==2.32.3
itsdangerous==2.2.0
Werkzeug==3.0.3
gunicorn==22.0.0
